+++ src/main/java/finalmission/woowabowling/lane/domain/LaneRepository.java
@@ -1,4 +1,4 @@
-package finalmission.woowabowling.lane;
+package finalmission.woowabowling.lane.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
