+++ src/main/java/finalmission/woowabowling/member/controller/MemberController.java
@@ -1,5 +1,7 @@
-package finalmission.woowabowling.member;
+package finalmission.woowabowling.member.controller;
 
+import finalmission.woowabowling.member.LoginRequest;
+import finalmission.woowabowling.member.service.MemberService;
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
