+++ src/main/java/finalmission/woowabowling/reservatoin/domain/ReservationRepository.java
@@ -1,4 +1,4 @@
-package finalmission.woowabowling.reservatoin;
+package finalmission.woowabowling.reservatoin.domain;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
